Removed 2.5 bc layer
