+++ src/Symfony/Component/Validator/Validator/RecursiveValidator.php
@@ -11,6 +11,9 @@
 
 namespace Symfony\Component\Validator\Validator;
 
+use Symfony\Component\Validator\Constraint;
+use Symfony\Component\Validator\Constraints\GroupSequence;
+use Symfony\Component\Validator\Constraints\Valid;
 use Symfony\Component\Validator\ConstraintValidatorFactoryInterface;
 use Symfony\Component\Validator\Context\ExecutionContextFactoryInterface;
 use Symfony\Component\Validator\Context\ExecutionContextInterface;
@@ -112,6 +115,22 @@ class RecursiveValidator implements ValidatorInterface
      */
     public function validate($value, $constraints = null, $groups = null)
     {
+        // The following code must be removed in 3.0
+        $numArgs = func_num_args();
+        if ($numArgs > 3) {
+            trigger_error('The '.__METHOD__.' method is deprecated in version 2.5 and will be removed in version 3.0. Use the Symfony\Component\Validator\Validator\ValidatorInterface::validate method instead.', E_USER_DEPRECATED);
+            $traverse = func_get_arg(3);
+            $deep = func_get_arg(4);
+            $constraints = new Valid(array('traverse' => $traverse, 'deep' => $deep));
+
+            if (self::testConstraints($groups)) {
+                if ((!$traverse && !$deep) || self::testGroups($traverse)) {
+                    $constraints = $groups;
+                    $groups = $traverse;
+                }
+            }
+        }
+
         return $this->startContext($value)
             ->validate($value, $constraints, $groups)
             ->getViolations();
@@ -136,5 +155,23 @@ class RecursiveValidator implements ValidatorInterface
         return $this->startContext(is_object($objectOrClass) ? $objectOrClass : $value)
             ->validatePropertyValue($objectOrClass, $propertyName, $value, $groups)
             ->getViolations();
+    }
+
+    /**
+     * @internal
+     * @deprecated since version 2.5, to be removed in 3.0.
+     */
+    private static function testConstraints($constraints)
+    {
+        return null === $constraints || $constraints instanceof Constraint || (is_array($constraints) && current($constraints) instanceof Constraint);
+    }
+
+    /**
+     * @internal
+     * @deprecated since version 2.5, to be removed in 3.0.
+     */
+    private static function testGroups($groups)
+    {
+        return null === $groups || is_string($groups) || $groups instanceof GroupSequence || (is_array($groups) && (is_string(current($groups)) || current($groups) instanceof GroupSequence));
     }
 }
